qualification types api JSON response

+++ Modules/Settings/Http/Controllers/QualilificationTypesController.php
@@ -8,20 +8,23 @@ use Illuminate\Routing\Controller;
 use Modules\Settings\Entities\QualificationTypes;
 use Modules\Settings\Http\Requests\QualificationTypeRequest;
 
-class QualilificationTypesController extends Controller {
+class QualilificationTypesController extends Controller
+{
 
-    public function __construct() {
-        $this->middleware(['permission:qualification-types_read'])->only('index');
-        $this->middleware(['permission:qualification-types_create'])->only('create');
-        $this->middleware(['permission:qualification-types_update'])->only('edit');
-        $this->middleware(['permission:qualification-types_delete'])->only('destroy');
+    public function __construct()
+    {
+//        $this->middleware(['permission:qualification-types_read'])->only('index');
+//        $this->middleware(['permission:qualification-types_create'])->only('create');
+//        $this->middleware(['permission:qualification-types_update'])->only('edit');
+//        $this->middleware(['permission:qualification-types_delete'])->only('destroy');
     }
 
     /**
      * Display a listing of the resource.
      * @return Response
      */
-    public function index() {
+    public function index()
+    {
         $query = QualificationTypes::query();
 
         if (request()->level_id > 0)
@@ -30,15 +33,16 @@ class QualilificationTypesController extends Controller {
         if (request()->academic_year_id > 0)
             $query->where('academic_year_id', request()->academic_year_id);
 
-        $qualification_types = $query->OrderBy('created_at', 'desc')->get();
-        return view('settings::qualification_types.index', compact('qualification_types'));
+        $qualification_types = $query->OrderBy('created_at', 'desc')->paginate(10);
+        return responseJson(1, "ok", $qualification_types);
     }
 
     /**
      * Show the form for creating a new resource.
      * @return Response
      */
-    public function create() {
+    public function create()
+    {
         return view('settings::qualification_types.create');
     }
 
@@ -47,19 +51,31 @@ class QualilificationTypesController extends Controller {
      * @param Request $request
      * @return Response
      */
-    public function store(QualificationTypeRequest $request) {
+    public function store(Request $request)
+    {
+        $validator = validator($request->all(), [
+            'name' => 'required',
+            'grade' => 'required',
+            'qualification_id' => 'required|exists:qualifications,id',
+            'academic_year_id' => 'required|exists:academic_years,id',
+            'percentage' => 'required',
+            'level_id' => 'required|exists:levels,id',
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
         try {
             $qualification_type = QualificationTypes::create($request->all());
             if ($qualification_type) {
-                notify()->success(__('data created successfully'), "", "bottomLeft");
-                return redirect()->route('qualification-types.index');
+                watch('qualification type was created', '');
+                return responseJson(1, __('data created successfully'), $qualification_type);
             } else {
                 notify()->error("هناك خطأ ما يرجى المحاولة فى وقت لاحق", "", "bottomLeft");
                 return redirect()->route('qualification-types.index');
             }
-        } catch (\Exception $th) {
-            notify()->error($th->getMessage(), "", "bottomLeft");
-            return redirect()->route('qualification-types.index');
+        } catch (\Exception $ex) {
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 
@@ -68,8 +84,13 @@ class QualilificationTypesController extends Controller {
      * @param int $id
      * @return Response
      */
-    public function show($id) {
-        return view('settings::show');
+    public function show($id)
+    {
+        $qualification_type = QualificationTypes::find($id);
+        if (!$qualification_type) {
+            return responseJson(0, __('data not found'), '');
+        }
+        return responseJson(1, "ok", $qualification_type);
     }
 
     /**
@@ -77,7 +98,8 @@ class QualilificationTypesController extends Controller {
      * @param int $id
      * @return Response
      */
-    public function edit($id) {
+    public function edit($id)
+    {
         $qualification = QualificationTypes::select()->find($id);
         if (!$qualification) {
 
@@ -94,20 +116,30 @@ class QualilificationTypesController extends Controller {
      * @param int $id
      * @return Response
      */
-    public function update(QualificationTypeRequest $request, $id) {
-        try {
-            $qualification = QualificationTypes::find($id);
-
-            if (!$qualification) {
-                notify()->warning(__('data not found'), "", "bottomLeft");
-            } else {
-                $qualification->update($request->all());
-                notify()->success(__('data updated successfully'), "", "bottomLeft");
-            }
-        } catch (\Exception $th) {
-            notify()->error($th->getMessage(), "", "bottomLeft");
+    public function update(Request $request, $id)
+    {
+        $qualification = QualificationTypes::find($id);
+        if (!$qualification) {
+            notify()->warning(__('data not found'), "", "bottomLeft");
         }
-        return redirect()->route('qualification-types.index');
+        $validator = validator($request->all(), [
+            'name' => 'required',
+            'grade' => 'required',
+            'qualification_id' => 'required|exists:qualifications,id',
+            'academic_year_id' => 'required|exists:academic_years,id',
+            'percentage' => 'required',
+            'level_id' => 'required|exists:levels,id',
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
+        try {
+            $qualification->update($request->all());
+            return responseJson(1, __('data updated successfully'), $qualification);
+        } catch (\Exception $ex) {
+            return responseJson(0, "", $ex->getMessage());
+        }
     }
 
     /**
@@ -115,19 +147,23 @@ class QualilificationTypesController extends Controller {
      * @param int $id
      * @return Response
      */
-    public function destroy($id) {
+    public function destroy($id)
+    {
         try {
             $setting = QualificationTypes::find($id);
             if (!$setting) {
-                notify()->warning(__('data not found'), "", "bottomLeft");
+                return responseJson(0, __('data not found'), '');
             } else {
+                $setting->applications->count();
+                if($setting->applications->count() > 0 || $setting->students->count() > 0)
+                    return responseJson(0, __('this item can not be deleted'), $setting->fresh());
+
                 $setting->delete();
-                notify()->success(__('data deleted successsfully'), "", "bottomLeft");
+                return responseJson(1, __('deleted successfully'), '');
             }
         } catch (\Exception $ex) {
-            notify()->error($th->getMessage(), "", "bottomLeft");
+            return responseJson(0, "", $ex->getMessage());
         }
-        return redirect()->route('qualification-types.index');
     }
 
 }
